lightning connected.
an options page with {endpoint, username, password}.

when the popup is loaded we show the default page, NOTHING,
which displays a crude list of transactions.

the popup calls the background page, which calls the spark
{endpoint} (or any endpoint that supports the same protocol),
which is something like http://your-ip:9737/rpc.

in this call we use the special header X-Access, made by an
arcane hmac function that uses {username} and {password}.